file_standardization_pipeline/requirements.txt
aws_ddk_core
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0